DataProcessforDQMC

Documentation for DataProcessforDQMC.

DataProcessforDQMC.format_energy

DataProcessforDQMC.format_pair_onsite_bulk

DataProcessforDQMC.format_pair_onsite_edge

DataProcessforDQMC.format_pair_onsite_k

DataProcessforDQMC.format_pair_onsite_r

DataProcessforDQMC.format_pair_onsite_r

DataProcessforDQMC.insert_parameters_cols

DataProcessforDQMC.read_format_data

DataProcessforDQMC.format_energy — Method

format_energy(data_dir::String)

Format the energy.bin in data_dir into a dataframe and store it in a .csv and .toml file.

DataProcessforDQMC.format_pair_onsite_bulk — Method

format_pair_onsite_bulk(data_dir::String, sigs::Array{Float64,1})

Format the pair_onsite_bulk.bin in data_dir into a dataframe and store it in a .csv and .toml file.

DataProcessforDQMC.format_pair_onsite_edge — Method

format_pair_onsite_edge(data_dir::String, sigs::Array{Float64,1})

Format the pair_onsite_edge.bin, pair_onsite_interedges.bin and pair_onsite_intraedges.bin in data_dir into a dataframe and store it in a .csv and .toml file.

DataProcessforDQMC.format_pair_onsite_k — Method

format_pair_onsite_k(data_dir::String, sigs::Array{Float64,1})

Format the pair_onsite_k.bin in data_dir into a dataframe and store it in a `.csv and .toml file. For now, only support one sublattice system.

DataProcessforDQMC.format_pair_onsite_r — Method

format_pair_onsite_r(data_dir::String, sigs::Array{Float64,1}, L::Int)

Format the pair_onsite_r.bin in data_dir into a dataframe and store it in a .csv and .toml file. For now, only support one sublattice system. This function support old-fashioned output data, i.e., only output the values without the coordinates, so users need to specify the lattice size L.

DataProcessforDQMC.format_pair_onsite_r — Method

format_pair_onsite_r(data_dir::String, sigs::Array{Float64,1})

Format the pair_onsite_r.bin in data_dir into a dataframe and store it in a .csv and .toml file. For now, only support one sublattice system. This function support new-fashioned output data, i.e., output both the coordinates and the values in each line.

DataProcessforDQMC.insert_parameters_cols — Method

insert_parameters_cols(df::DataFrame,data_dir::String)

Exact parameters from data_dir and insert them as columns into df. e.g. data_dir = ".../proj_majo_cy.b16.000.U-4.00.Delta_ppm0.4.mu-0.5.L8.dtau0.05/" will extract beta, U, Delta_ppm, mu, L and dtau from it.

DataProcessforDQMC.read_format_data — Method

read_format_data(dataname::String, data_dir::String)

Read dataname.csv and dataname.toml in data_dir.