DataProcessforDQMC

Documentation for DataProcessforDQMC.

DataProcessforDQMC.format_energy

DataProcessforDQMC.format_pair_onsite_edge

DataProcessforDQMC.format_energy — Method

format_energy(data_dir::String)

Format the energy.bin in data_dir into a dataframe and store it in a .csv and .toml file.

DataProcessforDQMC.format_pair_onsite_edge — Method

format_pair_onsite_edge(data_dir::String, sigs::Array{Float64,1})

Format the pair_onsite_edge.bin, pair_onsite_interedges.bin and pair_onsite_intraedges.bin in data_dir into a dataframe and store it in a .csv and .toml file.