DataProcessforDQMC

Documentation for DataProcessforDQMC.

DataProcessforDQMC.format_edgerdata

DataProcessforDQMC.format_edgerfulldata

DataProcessforDQMC.format_energy

DataProcessforDQMC.format_expRenyi_2p_edge

DataProcessforDQMC.format_kdata

DataProcessforDQMC.format_onecol

DataProcessforDQMC.format_pair_onsite_alledges

DataProcessforDQMC.format_pair_onsite_r

DataProcessforDQMC.format_rdata

DataProcessforDQMC.insert_parameters_cols

DataProcessforDQMC.read_format_data

DataProcessforDQMC.write_df

DataProcessforDQMC.format_edgerdata — Method

format_edgerdata(data_dir::String, obsname::String; ifsave::Bool = false)

A generic formatter for r-dependent data along the edge of a cylinder with each row corresponding to a bin and a spatial position. The spatial position is from 1 to L/2, where L is the length of the edge (typically an even number). Format the obsname.bin in data_dir into a dataframe and store it in a .csv and .toml file. In the .csv file, the first column is the bin index, the second column is the spatial position, and the third column is the value of the observable.

DataProcessforDQMC.format_edgerfulldata — Method

format_edgerfulldata(data_dir::String, obsname::String; ifsave::Bool = false)

A generic formatter for r-dependent data along the edge of a cylinder with each row corresponding to a bin and a spatial position. Contain all the two-point correlation information between all points along the same edges. Format the obsname.bin in data_dir into a dataframe and store it in a .csv and .toml file. In the .csv file, the first column is the bin index, the second column is the edge number, the third column is the starting point i, the fourth column is the distance from i, and the fifth column is the value of the observable.

DataProcessforDQMC.format_energy — Method

format_energy(data_dir::String, energy_list::Array{String,1})

Format the energy.bin in data_dir into a dataframe and store it in a .csv and .toml file.

DataProcessforDQMC.format_expRenyi_2p_edge — Method

format_expRenyi_2p_edge(data_dir::String)

Format the expRenyiN2_2p_edge.bin and expRenyiSth_2p_edge.bin in data_dir into a dataframe and store it in a .csv and .toml file.

DataProcessforDQMC.format_kdata — Method

format_kdata(data_dir::String, obsname::String; ifsave::Bool = false)

DataProcessforDQMC.format_onecol — Method

format_onecol(data_dir::String, obsname::String; ifsave::Bool = false)

A generic formatter for one-column data with each row corresponding to a bin. Format the obsname.bin in data_dir into a dataframe and store it in a .csv and .toml file. In the .csv file, the first column is the bin index, and the second column is the value of the observable.

DataProcessforDQMC.format_pair_onsite_alledges — Method

format_pair_onsite_alledges(data_dir::String)

Format the pair_onsite_alledges.bin, pair_onsite_interedges.bin and pair_onsite_intraedges.bin in data_dir into a dataframe and store it in a .csv and .toml file.

DataProcessforDQMC.format_pair_onsite_r — Method

format_pair_onsite_r(data_dir::String, L::Int)

Format the pair_onsite_r.bin in data_dir into a dataframe and store it in a .csv and .toml file. For now, only support one sublattice system. This function support old-fashioned output data, i.e., only output the values without the coordinates, so users need to specify the lattice size L. Just a compatibility function for old-fashioned output, and will be deprecated in the future.

DataProcessforDQMC.format_rdata — Method

format_rdata(data_dir::String, obsname::String; ifsave::Bool = false)

A generic formatter for r-dependent data with each row corresponding to a bin and a spatial position. Format the obsname.bin in data_dir into a dataframe and store it in a .csv and .toml file. For now, only support 2D system with one sublattice. In the .csv file, the first column is the bin index, the second and third columns are the x and y coordinates, and the fourth column is the value of the observable.

DataProcessforDQMC.insert_parameters_cols — Method

insert_parameters_cols(df::DataFrame,data_dir::String)

Exact parameters from data_dir and insert them as columns into df. e.g. data_dir = ".../proj_majo_cy.b16.000.U-4.00.Delta_ppm0.4.mu-0.5.L8.dtau0.05/" will extract beta, U, Delta_ppm, mu, L and dtau from it.

DataProcessforDQMC.read_format_data — Method

read_format_data(dataname::String, data_dir::String)

Read dataname.csv and dataname.toml in data_dir.

DataProcessforDQMC.write_df — Method

write_df(df::DataFrame, data_dir::String, dataname::String)

Write df into data_dir/formatted_data/dataname.csv and data_dir/formatted_data/dataname.toml.